Prevent quopri-encoding the Date header.
The Date header was ending up like this:

    Date: =?utf-8?q?Sat=2C_01_Sep_2012_13=3A08=3A29_-0300?=

I was opening the received message in Thunderbird, using my
gmail account over IMAP and the date was not being read
correctly (not sure if this is Thunderbird's fault) and the
message would appear as if it were from 1969...

Anyway, the problem goes away if this header is pre-converted
to a bytestring. I don't think there is ever anything other
than ASCII in this header anyway.

Thoughts welcome...